Updated ExportQ and CFM to use new ObserverProvider API
Created a new CombineQueue API that replaces the CollisionFreeMap.  The new
combineQ uses the new Fluo Observer APIs exclusively.  The CollisionFreeMap
only uses the old Fluo Observer APIs and was deprecated.

The ExportQueue was modified to support a Fluent configuration mechanism. This
can only be used when using the new Observer API.